+++ CHANGELOG.rst
@@ -5,6 +5,7 @@
 - Drop support for Python < 3.9.
 - Add tested support for Django 5.2 and Python 3.13.
 - Lift restriction on ``django.contrib.contenttypes`` being installed (#84).
+- Ensure ``UnsealedAttributeAccess`` propagates through ``getattr`` (#83).
 
 1.6.3
 =====